adding cat and ls functionalities

+++ libc/chdir.c
@@ -4,7 +4,7 @@
 #include<errno.h>
 int chdir(const char *path)
 {
-	int returnvalue =  syscall_1((uint64_t)SYS_chdir,(uint64_t)path);
+	uint64_t returnvalue =  syscall_1((uint64_t)SYS_chdir,(uint64_t)path);
     if(returnvalue < 0){
         errno = -(returnvalue);
         return -1;
