+++ libc/close.c
@@ -4,7 +4,7 @@
 #include<errno.h>
 int close(int fd)
 {
-	int returnvalue = syscall_1((uint64_t)SYS_close,(uint64_t)fd);
+	uint64_t returnvalue = syscall_1((uint64_t)SYS_close,(uint64_t)fd);
     if(returnvalue < 0){
         errno = -(returnvalue);
         return -1;
