+++ libc/dup.c
@@ -3,7 +3,7 @@
 #include<syscall.h>
 #include<errno.h>
 int dup(int oldfd){
-    int returnvalue = syscall_1(SYS_dup,oldfd);
+    uint64_t returnvalue = syscall_1(SYS_dup,oldfd);
     if(returnvalue < 0){
         errno = -(returnvalue);
         return -1;
