+++ libc/execve.c
@@ -4,7 +4,7 @@
 #include<syscall.h>
 int execve(const char *filename, char *const argv[], char *const envp[])
 {
-    int returnvalue = syscall_3((uint64_t)SYS_execve, (uint64_t) filename, (uint64_t) argv, (uint64_t) envp);
+    uint64_t returnvalue = syscall_3((uint64_t)SYS_execve, (uint64_t) filename, (uint64_t) argv, (uint64_t) envp);
     if(returnvalue < 0){
         errno = -(returnvalue);
         return -1;
